Add support for reporting analysis data to analytics
My primary goal from this review is to get feedback on whether this
looks like the right data to collect. Of course, any other feedback is
also appreciated.

Landing this will not cause any additional analytics to be collected,
but will allow us to locally explore the data that would be collected
in order to verify it's usefulness.

Change-Id: I2eaa2630673d181695d10f75d98e99f771a6f80c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/288380
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>