[analysis_server] Fix a race in LSP server when a file was opened and created at the same time
This would show up during a rename, where the server gets a textDocument/didOpen around the same time a file was created.

If things happened in this order:

1. File created on disk
2. Server processes textDocument/didOpen
3. Server processes watch event

Step 2 would skip analysis the overlay because it saw the file on disk and the contents are the same. Step 3 would skip analysis because there is an overlay that trumps the disk.

The fix is to ensure that Step 2 is using the contents from the servers state, and never looking at the disk.

Change-Id: I98c53e3cd4fbc6cd28208b9144596ea15de64394
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/291801
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>