Increase google builder priority.
The google builder takes 25 seconds to run and unblocks lkgr from being
able to find dev release candidates and advance the branch.
Unfortunately when the commit queue is massively overloaded with try
jobs of equal prioritiy, the google builder can be triggered for up to 6
hours and time out.

This change makes the google builder more important than try jobs and
shards, which has big benefits in improving lkgr rolls (the rest of the
lkgr builders run in their own fast pool), and is safe since it doesn't
block other work for more than a minute on a single bot.

Fixes: b/276851232
Change-Id: Ia87e8fa582c54ff0a473fead564f6bb26f6252dd
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/292723
Reviewed-by: Alexander Thomas <athom@google.com>
Commit-Queue: Jonas Termansen <sortie@google.com>