[web] Pin tests/web/regress/issue/230108748_test to 2.19.
The original bug suggests that the regression has to do with constant
evaluation in a switch. Given that 3.0 changes the way switches/patterns
are evaluated, I think it makes more sense to pin the test to 2.19 for
now than to find a valid reformulation of it.

Bug: #51676
Fixes: #51947
Change-Id: I7a6f4dee85624d3b9f2eec04fd9bb07310cdd310
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/293381
Reviewed-by: Sigmund Cherem <sigmund@google.com>
Commit-Queue: Mayank Patke <fishythefish@google.com>