[dart2js] Pass option to exclude source code from closed world phase output dill.
Dart2JS uses the source bytes to generate error messages for the CFE and during closed world construction (for const evaluation). But the actual source code is never used beyond that.

For large applications not including the source code can reduce the size of the dill by close to 50%.

Change-Id: I99d2cb5d1cea843bc12a5e90df3ed651ed8a1ff1
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/293762
Reviewed-by: Stephen Adams <sra@google.com>
Commit-Queue: Nate Biggs <natebiggs@google.com>
Reviewed-by: Sigmund Cherem <sigmund@google.com>