[dart2wasm] Use `identical` in switch statements
Currently when `switch` expression is nullable (so the values are boxed)
we currently use `ref.eq` to compare the values.

This doesn't work as expected when one of the values is a constant and
the other one is a runtime-allocated value. Example:

    void test(bool? x) {
      switch (x) {
        case false:
          print('no');
        case true:
          print('yes');
        case null:
          print('maybe');
      }
    }

    bool runtimeTrue = int.parse('1') == 1;

    void main() {
      test(runtimeTrue);
    }

Here the return value of `runtimeTrue` is boxed in the call site of
`test`. In the body of `test` we use globals for constants `true` and
`false`, which are never equal to runtime-allocated values.

We now use `identical` in these cases, which handles `bool`, `String`,
and `num` values as expected and compares the rest using `ref.eq`.

New passing test: co19/LanguageFeatures/Patterns/exhaustiveness_A01_t10

Fixes #52075

Change-Id: Ibbeda7525fd40cfec5ee477493d0d96ac6e139b2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/295941
Commit-Queue: Ömer Ağacan <omersa@google.com>
Reviewed-by: Aske Simon Christensen <askesc@google.com>