[CP][beta] Remove left-over patch declarations for `List` constructor.
This CL replaces https://dart-review.googlesource.com/c/sdk/+/296900

The `List` constructor is removed in Dart 3.0.
Some of the `@patch` implementations were not removed.

This is *high priority*. It seems the left-over `@patch factory List` constructor did not cause any errors, instead it *added* a constructor  to `List` that can be used in web compiled code. Even if `List` doesn't have such a constructor in the SDK code proper.
The VM and analyzer will say the invocation is an error, but dart2js happily compiles it and runs.

(It used to be that patches couldn't add public members, that security seems to have been removed.)

Also removes code which tries to detect "the unnamed List constructor",
which is no longer a thing, and a number of invocations of the constructor, where it's not clear that the test is aware that the constructor no longer exists, and is not marked as `@dart=2.x` with x < 12.

Tested: ci
Cherry-Pick: https://dart-review.googlesource.com/c/sdk/+/297100
Change-Id: I2340ab7f68be50fe22f9940bcaaf45c3ff6dc5ca
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/297461
Reviewed-by: Sigmund Cherem <sigmund@google.com>
Reviewed-by: Kevin Chisholm <kevinjchisholm@google.com>
Reviewed-by: Nate Bosch <nbosch@google.com>